+++ platform/exacto/commander/exacto_data_storage.h
@@ -116,7 +116,7 @@ extern uint8_t receiveExactoDataStorage();
 extern uint8_t setupReceiveLengthExactoDataStorage( const uint8_t length);
 extern uint8_t clearExactoDataStorage();
 extern thread_control_result_t getStateExactoDataStorage();
-extern uint8_t setDataToExactoDataStorage(uint8_t * data, const uint8_t datacount, thread_control_result_t result);
+extern uint8_t setDataToExactoDataStorage(uint8_t * data, const uint16_t datacount, thread_control_result_t result);
 extern uint8_t getMailFromExactoDataStorage(uint8_t * receiver, const uint8_t receiver_length);
 
 extern exactolink_package_result_t ex_checkData_ExDtStr();
