+++ platform/exacto/spi/spi_afina_mliner.c
@@ -40,7 +40,7 @@ static uint16_t SAM_PackageStart_Buffer = SAM_PACKAGE_START_POINTER;
 typedef struct
 {
     uint8_t dt_buffer[SPI1_FULL_DMA_RXTX_BUFFER_SIZE];
-    uint8_t dt_count;
+    uint16_t dt_count;
     struct mutex dt_mutex;
     struct lthread dt_lth;
     uint8_t is_full;
